requirements.txt
numpy
opencv-python
Pillow
tqdm
imageio
lmdb
torch 
torchvision